Move StudioMovieInfo.java from the test folder to the src folder.
